+++ CHANGELOG.md
@@ -36,9 +36,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add color styling to help output ([#387](https://github.com/wcampbell0x2a/backhand/pull/387))
 
 #### unsquashfs 
+- Changed name to `unsquashfs-backhand` ([#356](https://github.com/wcampbell0x2a/backhand/pull/356))
 - Add progress bar for a cleaner output when extracting files ([#272](https://github.com/wcampbell0x2a/backhand/pull/272))
 - Add `--quiet` for not displaying progress bar and RUST_LOG output ([#272](https://github.com/wcampbell0x2a/backhand/pull/272))
 - Add multiple threads for extracing files, giving us the same performance in most cases as `squashfs-tools/unsquashfs`! ([#278](https://github.com/wcampbell0x2a/backhand/pull/278))
+
+#### add
+- Changed name to `add-backhand` ([#356](https://github.com/wcampbell0x2a/backhand/pull/356))
+
+#### replace
+- Changed name to `replace-backhand` ([#356](https://github.com/wcampbell0x2a/backhand/pull/356))
 
 ### ci
 - Add testing and release binaries for the following platforms:([#259](https://github.com/wcampbell0x2a/backhand/pull/259))